% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDS-package.R
\docType{package}
\name{EDS-package}
\alias{EDS}
\alias{EDS-package}
\title{EDS: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/parpishro/EDS}
  \item Report bugs at \url{https://github.com/parpishro/EDS/issues}
}

}
\author{
\strong{Maintainer}: Parham Pishrobat \email{par.pishrobat@gmail.com} (\href{https://orcid.org/0009-0003-4422-8111}{ORCID}) [copyright holder]

}
\keyword{internal}
