% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Calculate the geometric mean of a numeric vector}
\usage{
geometric_mean(x, remove_NA = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{remove_NA}{logical scalar, indicating whether NA values should be
stripped before computation proceeds}
}
\value{
the geometric mean of the values in 'x', a numeric scalar value
}
\description{
Calculate the geometric mean of a numeric vector
}
\examples{
geometric_mean(x = 1:10)
geometric_mean(x = c(1:10, NA), remove_NA = TRUE)

}
