% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_local.R
\name{install_local}
\alias{install_local}
\title{Install a local development package to a custom location.}
\usage{
install_local(libPaths_element = 3, ...)
}
\arguments{
\item{libPaths_element}{Integer stating which path from \code{.libPaths()} to use as the installation location. Defaults to 3 based on assumed \code{.libPaths()} structure of (CRAN, Github, personal, base).}

\item{...}{Other arguments to pass to \code{devtools::install()}.}
}
\description{
A wrapper function for \code{devtools::install()}. Uses R CMD INSTALL to install the package to a location specified through \code{.libPaths()}.
}
