% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mean.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Calculate the rolling mean of a vector}
\usage{
rolling_mean(x, window_width, ...)
}
\arguments{
\item{x}{vector of values that can be interpreted numerically.}

\item{window_width}{integer specifying the number of values included in each mean calculation. Should be an odd, positive integer.}

\item{...}{Additional arguments to pass to the mean() function call.}
}
\value{
A vector of rolling mean values of the same length as x.
}
\description{
Calculate the rolling mean of a vector
}
\examples{
rolling_mean(x = 1:5, window_width = 3)
rolling_mean(x = rep(0:2, each = 6), window_width = 5)
rolling_mean(x = c(TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE), window_width = 3)

}
