% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eds.R
\name{run_eds}
\alias{run_eds}
\title{importFrom(stats, complete.cases, filter, quantile, var)
importFrom(utils, head, install.packages, read.csv, tail)}
\usage{
run_eds(lon, lat, unit, time, buffer = 0.25)
}
\description{
importFrom(stats, complete.cases, filter, quantile, var)
importFrom(utils, head, install.packages, read.csv, tail)
}
